+++ src/scripts/browser/utils/prefs-defaults.js
@@ -11,6 +11,7 @@ const defaults = {
   'launch-startup-hidden': true,
   'launch-quit': false,
   'links-in-browser': true,
+  'switch-workplace': false,
   'block-seen-typing': false,
   'close-with-esc': false,
   'quit-behaviour-taught': false,