+++ src/scripts/renderer/webview/listeners.js
@@ -37,7 +37,7 @@ webView.addEventListener('console-message', function (event) {
   const msg = event.message.replace(/%c/g, '');
   const fwNormal = 'font-weight: normal;';
   const fwBold = 'font-weight: bold;';
-  console.log('%cWV:%c ' + msg, fwBold, fwNormal);
+  console.log('WV: ' + msg, fwBold, fwNormal);
 });
 
 // Listen for title changes to update the badge
@@ -57,7 +57,7 @@ webView.addEventListener('page-title-updated', function () {
 
 // Handle url clicks
 webView.addEventListener('new-window', function (event) {
-  log('sending open-url', event.url);
+  log('sending open-url', event.frameName, event.url);
   ipcRenderer.send('open-url', event.url, event.options);
 });
 
@@ -114,9 +114,8 @@ webView.addEventListener('dom-ready', function () {
 // Listen for did-finish-load
 webView.addEventListener('did-finish-load', function () {
   // Remove top banner
-  webView.executeJavaScript("document.getElementsByClassName('_s15')[0].outerHTML = '';");
-  webView.setZoomLevel(1); webView.setZoomLevel(0); // Fix non-automatic resize
-  
+  webView.send('remove-top-banner');
+
   // Hide the loading splash screen
   const loadingSplashDiv = document.querySelector('.loader');
   loadingSplashDiv.style.opacity = 0;